refactor(blog): move app/_utils/ and lib/utils.ts into lib/utils/
The app/ folder is for routing-convention files only (page.ts,
layout.ts, route.ts, middleware.ts, etc.). Helpers don't belong
there even under a private '_'-prefixed folder.

This commit relocates:
  app/_utils/ui.ts     -> lib/utils/ui.ts
  app/_utils/format.ts -> lib/utils/format.ts
  lib/utils.ts (cn)    -> lib/utils/cn.ts

The lib/utils/ folder is the canonical home for browser-safe
helper functions grouped by concern: ui.ts (Tailwind helpers),
cn.ts (class-name merger), format.ts (string formatters).

Updates every consumer:
  - 7 page files import from ../lib/utils/ui.ts at correct depth
  - 6 components/ui/* import cn from ../../lib/utils/cn.ts
