+++ examples/blog/app/(marketing)/about/page.ts
@@ -1,5 +1,5 @@
 import { html } from '@webjskit/core';
-import { rubric, displayH1, codeChip } from '../../_utils/ui.ts';
+import { rubric, displayH1, codeChip } from '../../../lib/utils/ui.ts';
 
 export const metadata = { title: 'About: webjs blog' };
