+++ examples/blog/components/ui/alert.ts
@@ -24,7 +24,7 @@
  *
  * Design tokens used: --card, --card-foreground, --destructive, --muted-foreground.
  */
-import { cn } from '../../lib/utils.ts';
+import { cn } from '../../lib/utils/cn.ts';
 
 const BASE =
   'relative grid w-full grid-cols-[0_1fr] items-start gap-y-0.5 rounded-lg border px-4 py-3 text-sm has-[>svg]:grid-cols-[calc(var(--spacing)*4)_1fr] has-[>svg]:gap-x-3 [&>svg]:size-4 [&>svg]:translate-y-0.5 [&>svg]:text-current';
