+++ examples/blog/components/ui/button.ts
@@ -21,7 +21,7 @@
  * --secondary, --secondary-foreground, --accent, --accent-foreground,
  * --background, --input, --ring.
  */
-import { cn } from '../../lib/utils.ts';
+import { cn } from '../../lib/utils/cn.ts';
 
 const BASE =
   "inline-flex shrink-0 items-center justify-center gap-2 rounded-md text-sm font-medium whitespace-nowrap transition-all outline-none focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 disabled:pointer-events-none disabled:opacity-50 aria-invalid:border-destructive aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 [&_svg]:pointer-events-none [&_svg]:shrink-0 [&_svg:not([class*='size-'])]:size-4";
