+++ examples/blog/lib/utils/format.ts
@@ -1,7 +1,7 @@
 /**
- * Utilities used by routes inside app/. Lives under _utils/: any folder
- * whose name starts with `_` is private: the router ignores it completely,
- * so this file cannot accidentally become a route.
+ * App-wide browser-safe formatting helpers. Lives under lib/utils/
+ * (browser-safe helpers grouped by concern). app/ is reserved for
+ * routing-convention files only.
  */
 export function relativeTime(dateish: string | Date): string {
   const then = new Date(dateish).getTime();
